+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Fixed External Storage not displaying empty allowlist warning.
 -   Fixed incrementing starting from 1 in amount screens not having an intended off-by-one.
 -   Fixed problems moving network devices with "Carry On" mod.
+-   Fixed escape key not working on auto-selected Grid search box.
 
 ## [2.0.0-milestone.3.3] - 2024-02-17